Add Model and Integration
Added Model for JobApplication, and ApplicationManager. Integrated a main(duke) that talks to UI, which gives the input the parser. Parser is not yet implemented, but already added commands and logic to take the parsed command and respectively execute correctly on the application list. Addtionally also integrated Ui table printing with ApplicationManager